==> 01c-unzip-and-split.ipynb <==
# Unzip and Split

M1 APPLIED ECONOMETRICS, Spring 2024

Applied Econometrics - Master TSE 1 - 2023/2024

> Sunlight Synchronization: Exploring the Influence of Daylight Saving Time on
> CO2 Emissions and Electricity Consumption in Australia's Electricity Grid

LAST MODIFIED: 29/02/2024 

LAST MODIFIED BY: Matthew Davis

Script duration: 18 hours (with SSD and multiprocessing). With external hard drive and single processing, could take >1 week.

Disk storage requirement: 200GB, in addition to the large requirements of adjacent scripts

Bandwidth requirement: None. No internet connectivity required.

Memory requirement: negligible. Any modern laptop/desktop will be sufficient.

--------------------

## Unzip

In a previous script we downloaded a lot of `.zip` files (and some `.csv`). Now we want to unzip them.
Note that some files are zips of zips of CSVs. Some zips have hundreds of files inside them (which I call "children" or "members").

The unzipped files are too large to fit onto my hard drive. (approx 2.4TB). So we compress each file with gzip (`.csv.gz`). Note that most tools (e.g. R, pandas, Excel) can decompress that on the fly.
It may seem pointless to decompress and immediately recompress. The aim is to split up the zips which contain many .csv files, and un-nest the recursive zipping. We also discard a lot of data (described more below.) and split each CSV file.

## Convert (Split) Files

After unzipping files, we're left with what look like a bunch of CSVs. But if you open them up you'll see they're not a normal CSV. They're a concatenation of many CSV files into one. (e.g. note how the number of columns changes as you scroll down.) So we want to unconcatenate them, or "split" them. So each ".CSV" file becomes many ".CSV" files.

To understand this format, read `Guide to CSV Data Format Standard.pdf`, which is saved in `./documentation`. Available [here](https://web.archive.org/web/20230414160249/https://aemo.com.au/-/media/files/market-it-systems/guide-to-csv-data-format-standard.pdf?la=en).

To read these files, look at the first column, which is C, D, or I.

* `C` means this is metadata. Typically the first and last row of the file. The first row tells you things like when the file was generated. We can probably ignore this. (Other timestamps are inside the data itself.)
* `I` means this is the header row for a new table
* `D` means this is a data row for the same table as the previous row

In the `D` and `I` rows, the second and third column contain info about which table this data belongs to (`REPORT_NAME` and `REPORT_SUBTYPE`). This is described more below. For now we put output files under a 2-layer folder structure based on this. 

The fourth column is an integer related to some kind of schema versioning. We'll keep that, so that later if we find duplicate/clashing rows, choose the higher version number when duplicating.

As an example, consider this fictitious file:

```
C,SETP.WORLD,TER_DAILY,AEMO,PUBLIC,2024/01/07,08:00:01,0000000407795252,,0000000407795252
I,A,B,123,"col1","col2"
D,A,B,123,a1,b1
D,A,B,123,a2,b2
I,C,E,321,col_alpha,col_beta,col_sigma
D,C,E,321,x1,x1,x1
D,C,E,321,y2,y2,y2
D,C,E,321,z3,z3,z3
C,"END OF REPORT",9
```

Which contains two dataframes:

```
"col1","col2"
a1,b1
a2,b2
```
```
col_alpha,col_beta,col_sigma
x1,x1,x1
y2,y2,y2
z3,z3,z3
```

So for each input file, we want multiple output files. I call this process 'splitting'.

Note that pandas cannot read these original files, because the datatype changes across rows within the same file, and the number of rows changes. But the `csv` library can handle this.

## Map each file chunk to a table

In the sample file above, we split one file into two. We discarded the first few columns, with values `C,E` and `A,B`. These columns tell us what dataset each chunk belongs to. I call them "tables" because AEMO expects this data to end up in a SQL database.

These columns are called `REPORT_TYPE` and `REPORT_SUBTYPE`. We need to map these to a table name.
Sometimes it's obvious. E.g. `DISPATCH` and `PRICE` belongs to the `DISPATCHPRICE` table. But sometimes it's not obvious. (And sometimes there are blanks.) For each table, there are multiple pairs of `REPORT_TYPE` and `REPORT_SUBTYPE` for that table. This is not properly documented. The best way I've found is to look up those two values in a particular file. That file is no longer on the internet. But we can find an older version [here](https://web.archive.org/web/20220812083311/https://visualisations.aemo.com.au/aemo/di-help/Content/Data_Model/MMS_Data_Model.htm). (That's for schema version 5.1. We're up to 5.2 now. But the differences are small enough that it shouldn't be an issue. We care about so few tables that we could probably hard-code this mapping.)

There are a lot of exceptions which I have hard coded.

Note that for each output file, we save it with hive style partitioning.
In the example file above, see the values `123` and `321`. Those are versions of the schema for each table. They almost never matter. But for one table (I forget which) they do. So we want to keep that, and use it when deduplicating later. (For clashing rows, choose the more recent schema version.)
There's also a timestamp in the first row, which I call the "top timestamp". Again we want to keep this to help us decide which row to keep when deduplicating later.
Since both of these values are the same for every row in the output file, it would be wasteful to write them into the file itself. So we put them in the folder name. This is a standard practice called "Hive partitioning". e.g. if the second chunk above, with values `C,E` maps to table `sometable`, and the original file was `somefile.csv`, then the output file will be `something/sometable/SCHEMA_VERSION=321/TOP_TIMESTAMP=2024_01_07_08_00_01/somefile.csv.gz`. Many tools (e.g. `read_dataset` in R) will read these values automatically when you point them at folder `something/sometable/`.

## Combining together

For performance reasons, we do all 3 steps at once, including recompressing the final output to `.csv.gz`.
This makes it a bit confusing to debug and reason with. But it speeds up processing by not writing out the intermediate `.csv` files to disk. (And saves hundreds of GB of disk space.)

To help debug this, files we can't process (e.g. corrupt zips, or CSVs that don't match the expected format) are copied to an `error` folder. (Because the file might not exist directly in the source directory, if it's unzipped from an unzipped zip.)

When handling errors, we try not to let an error in file 20000 stop us from attempting to process files 20001, 20002 etc. So we wrap each attempt in a `try`/`except` block. We also delete any half-written files if there's an error. 

Watch out, there are two particular zip files which are corrupt. (They weren't downloaded wrong, they were just published as corrupt files.) I've checked those two and they're not relevant for us. So when you run this script, you will get an error that some files haven't been processed. If it's only a handful, check each on a case-by-case basis. (The errors are saved in text files next to the troublesome files, in the error directory.) And then decide to ignore them (or not).

## Performance

This script takes a long time. Multiprocessing makes it faster. Set `use_multiprocessing=True` to do that. 

## Imports

%pip install -r requirements.txt

import zipfile
import os
import shutil
from typing import List, Dict, Tuple, Optional, TextIO, BinaryIO
import csv
import traceback
import json
from random import shuffle
import re
from uuid import uuid4
import io
import gzip
import importlib
import sys

from tqdm import tqdm # progress bar animation
import pandas as pd

# There's a bug with multiprocessing in Jupyter on some OS'
# https://stackoverflow.com/a/65001152
# use drop-in replacement
#from multiprocessing import Pool
from multiprocess import Pool


# utils is our local utility module
# if we change utils.py, and re-run a normal 'import'
# python won't reload it by default. (Since it's already loaded.)
# So we force a reload
import utils
importlib.reload(utils)

assert sys.version_info >= (3, 6), "Python version too low."

## Path constants

You may need to change these, if you're trying to rerun the code on your own machine.

# all the large files will go here
# you need a few hundred GB of space
# this should contain the results of the previous scripts
base_data_dir = 'data'

# the results of the unzipping script were saved here
source_dir = os.path.join(base_data_dir, '01-A-raw')

# For each file in source_dir, after processing it will be moved here
# (so if we stop and restart processing, we don't reprocess files)
# only if move_when_done=True
archive_dir = os.path.join(base_data_dir, '01-C-raw-done')
temp_dir = os.path.join(base_data_dir, '01-C-split-mapped-csv-temp/')
dest_dir = os.path.join(base_data_dir, '01-C-split-mapped-csv/')

debug_path = os.path.join(base_data_dir, 'debug/')
bad_zip_dir = os.path.join(base_data_dir, 'error/bad_zip/')
bad_csv_dir = os.path.join(base_data_dir, 'error/bad_csv/')
log_file = os.path.join(base_data_dir, 'logs/01c.txt')
schemas_path = os.path.join(base_data_dir, '01-aemo-schemas.json')

metadata_path = os.path.join(base_data_dir, 'raw/MMS_Data_Model_Table_to_File_to_Report_Relationships_51.xlsx')

# if you want to process all tables (other than ones excluded further below) leave this empty
# if you specify only some tables in this list, only they will be processed
only_tables = []

## Other constants

You'll rarely need to change these.

# when finished processing a zip
# move it to archive_dir
# so if we stop and restart the playbook, we don't redo that zip
# set to True when you think the code is good.
# set to False when still writing the code and testing,
# or if you've changed any of the code.
move_when_done = False


# process files with multiprocessing, to be faster.
# If set to False, will use a for loop, which gives clearer traceback error messages.
# We found on some laptops multiprocessing didn't work because of Python installation issues.
# Defaulting to False to ensure compatability. 
# Set this to True if you know your python installation can do multiprocessing.
# This script will leave one CPU spare so you can keep using your laptop for other stuff.
use_multiprocessing = False


# how hardcore the gzip algorithm is for .csv.gz files. Low to prioritise speed over size
# but if you have an external hard drive, too low is probably slow and large
# (it's about disk write speed vs CPU compression speed) 
compresslevel = 3

version_col_name = 'SCHEMA_VERSION'
top_timestamp_col_name = 'TOP_TIMESTAMP'

# these are for what's inside metadata_path
sheet = 'MMSDM_v5-1'
report_subtype_column = 'PDR_REPORT_SUB_TYPE'
report_name_column = 'PDR_REPORT_NAME'
table_name_column = 'MMSDM_TABLE_NAME'

REPORT_NAME_NULL_PLACEHOLDER = 'NULL'
REPORT_SUBTYPE_NULL_PLACEHOLDER = 'NULL'

# AEMO data is always in UTC+10
TIMEZONE_OFFSET=10

# I've found some particular files have funny encodings
# just ignore them. We don't need them for our analysis.
csvgz_files_to_ignore = [
    # regex patterns
    r"7_days_High_Impact_Outages_\d+.csv.gz",
    r"High_Impact_Outages_\d+",
]

# some files hosted by AEMO are just corrupt
# We can't do anything about it. Thankfully these aren't relevant for our study.
known_bad_zips = [
    "PUBLIC_DISPATCHPRICES_PRE_AP_20230820.zip", 
    "PUBLIC_DISPATCHPRICES_PRE_AP_202308200745_0000000395024647.zip",
    "PUBLIC_TRADING_CUMULATIVE_PRICE_20230820_20230826.zip",
    "PUBLIC_TRADING_CUMULATIVE_PRICE_202308200745_0000000395024640.zip",
]

# Stuff that should go in a shared library

# decrease the priority of this script given by the OS
# so you can keep using your laptop for other stuff
utils.renice()

logger = utils.Logger(log_file)
logger.info("Initialising Logger")

## Unzip

For now we just define functions, but don't actually execute them.

for d in [temp_dir, dest_dir, archive_dir]:
    utils.create_dir(d)

# call process()
# catch (almost) all errors, and return them
# then after we process all files, check these errors
def _process_raw(path) -> Tuple[Optional[int], Optional[Exception]]:
    try:
        process_raw(path)
    except KeyboardInterrupt:
        raise
    except Exception as ex: # catch anything except keyboard interrupt
        logger.error(f"Failed to unzip {path}: {ex}")
        return ex
    else:
        if move_when_done:
            # move the file to another folder
            # so if we stop and restart, we will not 
            archive_path = os.path.join(archive_dir, os.path.basename(path))
            os.rename(path, archive_path)
        return None

# takes a string of a path to a file on disk
# either .csv or .zip
def process_raw(path: str):
    if path.lower().endswith('.csv'):
        with open(path, 'r', newline='') as f:
            process_csv([path], f)
    elif path.lower().endswith('.zip'):
        with open(path, 'rb') as f:
            process_zip([path], f)
    else:
        logger.warning(f"Ignoring file {path}")
    #logger.debug(f"Finished processing {path}")

# takes in a list of filenames
# e.g. if there's a/myzip.zip on disk, which contains otherzip.zip
# for the inner zip paths is ['a/myzip.zip', 'otherzip.zip']
# These names are just for debugging and logging 
# the file-like object f is used for reading content
def process_zip(fnames: List[str], f: BinaryIO):
    if isinstance(fnames, str):
        fnames = [fnames]
        
    try:
        with zipfile.ZipFile(f) as z:
            exceptions = []
            for m in z.namelist():
                try:
                    if m.lower().endswith('.zip'):
                        with z.open(m, mode='r') as mf:
                            process_zip(fnames + [m], mf)
                    elif m.lower().endswith('.csv'):
                            with z.open(m, mode='r') as mfb:
                                with io.TextIOWrapper(mfb, newline='') as mft:
                                    process_csv(fnames + [m], mft)
                    else:
                        paths_debug = ' -> '.join(fnames)
                        logger.warning(f"Found unknown file {m} in zip {paths_debug}")
                
                except KeyboardInterrupt:
                    raise
                except Exception as ex:
                    # ignore this one for now,
                    # continue processing the other files in this zip
                    # then rethrow this later
                    exceptions.append(ex)
    except zipfile.BadZipFile as e:
        f.seek(0) # go back to the start of the file
        debug_path = os.path.join(bad_zip_dir, os.path.basename(fnames[-1]))
        if not os.path.exists(os.path.dirname(debug_path)):
            try:
                os.makedirs(os.path.dirname(debug_path))
            except FileExistsError:
                # race condition when multiprocessing
                pass
        with open(debug_path, 'wb') as f_out:
            shutil.copyfileobj(f, f_out)
            
        # save the error to a file
        with open(debug_path + '.error.txt', 'w') as f_err:
            f_err.write(f"Issue is in file heirachy: {fnames}\n")
            f_err.write(str(e) + "\n")
            traceback.print_exc(file=f_err)
        zip_name = fnames[-1].split('/')[-1]
        if zip_name in known_bad_zips:
            logger.warning(f"Zip {zip_name} is bad. But we expected this particular one to be bad. Continuing")
        else:
            raise
    if exceptions:
        # rethrow an exception from processing one file in this zip
        raise exceptions[0]

## Split

### Atomic writes

It's possible to have the same final file be present in multiple source files. Since we're using multiprocessing, this could lead to race conditions where two processes write to the same file at once. That will jumple stuff up and corrupt the file. So we want some kind of atomicity.

To achieve this, we write to a random file path on the same disk. Then once it's done, we move that file across. We assume that moving a file to another folder on the same physical disk is atomic. 

I wanted to subclass `ioTextIOBase`, but I can't figure out how. Some attributes can't be overwritten. It's quite confusing. So I just wrote a vanilla class with the standard methods.

This class can handle compressed `.csv.gz` or uncompressed data.

class AtomicFile: #(io.TextIOBase):
    # newline='' is for csv.writer
    def __init__(self, temp_path, final_path, newline=''):
        self.temp_path = temp_path
        self.final_path = final_path
        
        # create destination folder if it doesn't exist
        if not os.path.exists(os.path.dirname(temp_path)):
            try:
                os.makedirs(os.path.dirname(temp_path))
            except FileExistsError:
                # race condition when multiprocessing
                pass
        #logger.info(f"Openning {temp_path} for writing")
        if temp_path.lower().endswith('.gz'):
            self.temp_f = gzip.open(temp_path, mode='wt', compresslevel=compresslevel, newline=newline)
        else:
            self.temp_f = open(temp_path, mode='w', newline=newline)
        self.closed = False

    def read(self, size=-1):
        assert not self.closed
        return self.temp_f.read(size)

    def write(self, s):
        assert not self.closed
        return self.temp_f.write(s)

    def close(self):
        
        if not self.closed:
            #logger.info(f"Closing {self.temp_path} after writing")
            self.temp_f.close()
            self.closed = True

            if not os.path.exists(os.path.dirname(self.final_path)):
                try:
                    os.makedirs(os.path.dirname(self.final_path))
                except FileExistsError:
                    # race condition when multiprocessing
                    pass
            
            os.rename(self.temp_path, self.final_path)
            #logger.info(f"Moving {self.temp_path} to {self.final_path}")
            self.delete_temp_dir()

    def delete_temp_file(self):
        try:
            os.remove(self.temp_path)
        except FileNotFoundError:
            pass

    def delete_final_file(self):
        try:
            os.remove(self.final_path)
        except FileNotFoundError:
            pass
        
    def delete_temp_dir(self):
            # in our case, the parent directory is likely now empty
            # delete it, if there are no other files in it
            try:
                os.rmdir(os.path.dirname(self.temp_path))
            except OSError:
                # another file is in the same directory
                # (unlikely, given how we partition the directory)
                pass   
                
    # like close, but delete everything
    def abort(self):
        
        
        if not self.closed:
            self.temp_f.close()
            self.closed = True
            self.delete_temp_file()
            self.delete_temp_dir()

    def tell(self):
        return self.temp_f.tell()


# the largest tables (larger than the smallest 300 tables combined) half a lot of columns we don't care about
# drop them in this script, to save time/disk space.
# columns_to_skip(table) tells you which columns to drop

with open(schemas_path, 'r') as f:
    schemas = json.load(f)
    
def get_columns_to_drop(table):
    return schemas[table].get('columns_to_drop', [])

assert get_columns_to_drop('DISPATCHREGIONSUM')
assert get_columns_to_drop('DISPATCHLOAD')

schemas.keys()


# takes a CSV
# fnames is a list of files
# first is the raw file on disk
# last is the CSV itself
# e.g. if this is a.txt inside b.zip inside c.zip
# fnames is ['c.zip', 'b.zip', 'a.csv']
# 
# we write to a random filename, then move to the destination
# so that if two processes are unzipping different files to the same destination concurrently, they won't corrupt the file.
#
# returns number of bytes written (uncompressed)
def process_csv(fnames: List[str], f_in: TextIO) -> int:
    if isinstance(fnames, str):
        fnames = [fnames]
    csv_name = os.path.basename(fnames[-1])

    try:
        csv_r = csv.reader(f_in)
        first_row = next(csv_r)
        assert isinstance(first_row[0], str), f"First row is wrong type, probably bytes not string. Is {type(first_row[0])}"
        # this keeps track of the 'row'
        # but note that some data values are text including a newline character
        # this is the number of records in the spreadsheet, not lines of text
        record_num = 1
        expected_start = 'C,SETP.WORLD'
        if first_row[0] != 'C':
            # some other files end up in the dataset
            # e.g. int668_v1_schedule_log_rpt_1~20231124133149.csv.gz
            # just ignore files like that
            logger.warning(f"First line does not start with C, {first_row[:10]=} in {fnames}. Ignoring")
            return

        # the first row contains a timestamp for when the file was created
        # e.g. 2020/06/04,00:03:11 (as two CSV cells)
        # change that to 2020_06_04_00_03_11
        # (only characters that can go in a folder name)
        t1 = first_row[5]
        if len(first_row) >= 6+1:
            t2 = first_row[6]
            top_timestamp_s = t1 + '_' + t2
        else:
            top_timestamp_s = t1
        top_timestamp_s = top_timestamp_s.replace('/', '_').replace(' ', '_').replace(':', '_').replace('\\', '_')
        assert '/' not in top_timestamp_s, f"Bad top timestabl: {first_row=}"
        
        row = next(csv_r)
        record_num = record_num + 1
    
        # these will be set later
        csv_w = None
        chars_to_skip = None
        f_out = None
        cols_to_drop = []
        
        while True:
            if row[0] in ['C', 'I']:
                # close off the last output file
                csv_w = None 
                if f_out is not None:
                    f_out.close()

                    if num_d_rows == 0:
                        logger.warning(f"No data rows, only header row written for {fnames}, so deleting output file")
                        f_out.delete_final_file()
                    
                
            if row[0] == 'C':

                # 2nd C line, which should be the last line of the file
                if row[1] == 'END OF REPORT':
                                                    
                    assert f_in.read().strip() == '', "Remainder after end of file"

                    checksum = int(row[-1])
                    if csv_name != 'PUBLIC_PDR_CONFIG_non_mms_data_model.CSV':
                        # sometimes AEMO do the checksum wrong, and there's a -1 there
                        assert checksum in (record_num, record_num - 1), f"Checksum on last line doesn't match in {fnames}, {checksum=} {record_num=}"
        
                    break
                elif row[1] == 'SETTLEMENTS RESIDUE CONTRACT REPORT':
                    logger.info(f"Ignoring C line {csv_r.line_num} in {fnames}: {row[1]}")
                else:
                    # additional C lines can be used for arbitrary comments
                    logger.warning(f"Unexpected C line {csv_r.line_num} in {fnames}: {row}")
            elif row[0] == 'I':
                # start of new file
                # The 'version' value is an int even on I rows
                # let's write a column header instead for this I row
                # then include the version value in subsequent D rows
                # choosing a column name that won't clash with others
                row_type, report_name, report_subtype, version, *remainder = row
                cols_to_skip = 4
    
                if report_name in ['', None]:
                    report_name = REPORT_NAME_NULL_PLACEHOLDER
                if report_subtype in ['', None]:
                    report_subtype = REPORT_SUBTYPE_NULL_PLACEHOLDER

                # which SQL 'table' does this map to?
                # e.g. 'DISPATCH', 'PRICE' maps to 'DISPATCHPRICE'
                table = map_table(report_name, report_subtype)

                if (table is None) or (table in tables_to_skip) or (only_tables and table not in only_tables):
                    if (table is None) and ((report_name, report_subtype) not in packages_to_ignore):
                        logger.warning(f"Skipping/ignoring {report_name}, {report_subtype} -> {table} in {fnames} with columns {remainder}")
                    dest_folder = None
                    final_path = None
                    dst = None
                    skip = True
                    assert (f_out is None) or f_out.closed
                    f_out = None
                    cols_to_drop = []
                else:

    
                    # for the file, include two columns in the folder, not the file itself
                    # i.e. hive-style partitioning
                    # (this saves space compared to repeating it on each line.)
                    # we use these values for deduplicating rows across files.
                    subdir = os.path.join(table, f"{version_col_name}={version}", f"{top_timestamp_col_name}={top_timestamp_s}")

                    final_path = os.path.join(dest_dir, subdir, csv_name + '.gz')
                    temp_path = os.path.join(temp_dir, subdir, csv_name + '.' + str(uuid4()) + '.tmp.gz')
                    
                    f_out = AtomicFile(temp_path, final_path)
                    #f_out = open(temp_path, 'w')
                    csv_w = csv.writer(f_out)

                    input_col_names = remainder
                    cols_to_drop = get_columns_to_drop(table)
                    output_col_names = [c for c in input_col_names if c not in cols_to_drop]
                    csv_w.writerow(output_col_names)
                    skip = False
                
                num_d_rows = 0
    
            else:
                assert row[0] == 'D', f"Unexpected non-D row: {row[:100]}" 
                if not skip:
                    data = row[cols_to_skip:]
                    if cols_to_drop:
                        data = [d for (c,d) in zip(input_col_names, data) if c in output_col_names]
                    csv_w.writerow(data)
                num_d_rows += 1
    
            row = next(csv_r)
            record_num = record_num + 1
    except KeyboardInterrupt:
        raise
    except Exception as e: 
        # close the output file if open
        try:
            if f_out is not None:
                f_out.abort()
        except NameError:
            pass
        try:
            os.remove(temp_path)
        except (FileNotFoundError, NameError, OSError):
            pass
        if any(re.search(ptn, csv_name) for ptn in csvgz_files_to_ignore):
            logger.info(f"Ignoring error splitting {csv_name}")
            return
        else:
            # save the file itself (since it's inside a zip, it's annoying to find for debugging)
            f_in.seek(0) # go back to the start of the file
            debug_path = os.path.join(bad_csv_dir, csv_name) + '.gz'
            if not os.path.exists(os.path.dirname(debug_path)):
                try:
                    os.makedirs(os.path.dirname(debug_path))
                except FileExistsError:
                    # race condition when multiprocessing
                    pass
            with gzip.open(debug_path, 'wt') as f_out_copy:
                s = f_in.read(1)
                assert isinstance(s, str), f"s is {type(s)}"
                f_in.seek(0)
                shutil.copyfileobj(f_in, f_out_copy)
                
            with open(debug_path + '.error.txt', 'w') as f_err:
                f_err.write(f"Issue is in file heirachy: {fnames}\n")
                try:
                    f_err.write(f"On text line {csv_r.line_num}")
                except NameError:
                    # csv_r is not defined
                    pass
                try:
                    f_err.write(f"CSV record {record_num}")
                except NameError:
                    pass
                f_err.write(str(e) + '\n')
                traceback.print_exc(file=f_err)
            raise

## Map

metadata_df = pd.read_excel(metadata_path, sheet_name=sheet)

metadata_df[report_name_column] = metadata_df[report_name_column].fillna(REPORT_NAME_NULL_PLACEHOLDER)
metadata_df[report_subtype_column] = metadata_df[report_subtype_column].fillna(REPORT_SUBTYPE_NULL_PLACEHOLDER)




# there are some packages I do not know how to map to tables
packages_to_ignore = [
    # maybe there's a table for these,
    # or they're legacy tables
    # these are ones we don't care about
    ('DINT', REPORT_SUBTYPE_NULL_PLACEHOLDER),
    ('TINT', REPORT_SUBTYPE_NULL_PLACEHOLDER),
    ('DCONS', REPORT_SUBTYPE_NULL_PLACEHOLDER),
    ('DREGION', REPORT_SUBTYPE_NULL_PLACEHOLDER),
    ('SPDCPC', REPORT_SUBTYPE_NULL_PLACEHOLDER),
    ('SRAFINANCIALS', 'RECONCILIATION_SUMMARY'),
    ('DAILY', 'MLF'), # electrical transmission loss factors.
    ('BILLING_CONFIG', 'BILLSMELTERRATE'), # alumninium smelter info. Possibly relevant? (Smeltering makes up 30% of NSW load)
    ('BILLING', 'CSP_SUPPORTDATA_SRA'),
    ('BILLING', 'ASPAYMENT_SUMMARY'), # probably belongs to BILLINGASPAYMENTS table, but this is not relevant to us so I haven't bothered checking
    ('BILLING', 'DIRECTION_CRA'),
    ('TRADING', 'CUMULATIVE_PRICE'),
    ('TREGION', 'NULL'),
    ('DAILY', 'WDR_NO_SCADA'),
    ('METER_DATA', 'GEN_DUID'),
    ('SEVENDAYOUTLOOK', 'PEAK'),
    ('TUNIT', 'NULL'),
    ('GPG', 'MARKET_SUMMARY'),
    ('GPG', 'CASESOLUTION'),
    ('GPG', 'CONSTRAINTSOLUTION'),
    ('GPG', 'PRICESOLUTION'),
    ('GPG', 'INTERCONNECTORSOLUTION'),
    ('CAUSER_PAYS_SCADA', 'NETWORK'),
    ('PDR_REPORT', 'COLUMN'),
    ('DUNIT', 'NULL'),
    ('YESTBID', 'BIDDAYOFFER'),
    ('YESTBID', 'BIDPEROFFER'),
    ('RESIDUE_PRICE_OFFER', 'NULL'),
    ('RESIDUE_PRICE_BID', 'NULL'), # columns don't match the RESIDUE_PRICE_BID table, almost match RESIDUE_PRICE_FUNDS_BID
    ('IBEI', 'PUBLISHING'),
    ('EMSLIMITS', 'LIM_ALTLIM'),
    ('DEMAND', 'HISTORIC'),

    
    ('PDR_REPORT', 'TABLE'),    # manifest metadata
    ('PDR_REPORT', 'FILE'),    # manifest metadata
    ('PDR_REPORT', 'MAPPING'),    # manifest metadata
    ('PDR_REPORT', 'COLUMN'),    # manifest metadata

    # Gas data
    ('GSH', 'PARTICIPANT_OPSTATE'),
    ('GSH', 'PARTICIPANTS'),
    ('GSH', 'AUCTION_CURTAILMENT_NOTICE'),
    ('GSH', 'AUCTION_PRICE_VOLUME'),
    ('GSH', 'BENCHMARK_PRICE'),
    ('GSH', 'PARK_SERVICES'),
    ('GSH', 'AUCTION_QUANTITIES'),
    ('GSH', 'FACILITIES'),
    ('GSH', 'TRANSACTION_SUMMARY'),
    ('GSH', 'HISTORICAL_SUMMARY'),
    ('GSH', 'NOTIONAL_POINTS'),
    ('GSH', 'REVISED_AUCTION_QUANTITIES'),
    ('GSH', 'PIPELINE_SEGMENTS'),
    ('GSH', 'CAPACITY_TRANSACTION'),
    ('GSH', 'ZONES'),
    ('GSH', 'SERVICE_POINTS'),

    # we should maybe double check
    ('FORCE_MAJEURE', 'MARKETSUSREGION'),
    ('FORCE_MAJEURE', 'MARKETSUSPENSION')

]


# there are some packages which aren't in the spreadsheet, but we can guess them
# in particular there's a CO2 one I can't 
package_exceptions = {
    ('DISPATCH', 'CASESOLUTION'): 'DISPATCHCASESOLUTION',
    ('GENCONSETTRK', REPORT_SUBTYPE_NULL_PLACEHOLDER): 'GENCONSETTRK',
    ('DISPATCH', 'REGIONFCASREQUIREMENT'): 'DISPATCH_FCAS_REQ',

    # this one is possibly useful
    # example file is CO2EII_AVAILABLE_GENERATORS.CSV.gz
    # It looks like it might go into BILLING_CO2E_PUBLICATION, but the columns are different
    # Example:
    #STATIONNAME,DUID,GENSETID,REGIONID,CO2E_EMISSIONS_FACTOR,CO2E_ENERGY_SOURCE,CO2E_DATA_SOURCE
    #"Appin Power Plant",APPIN,APPIN,NSW1,0.56318004,"Coal seam methane",NGA2022
    # so define a new table for it
    ('CO2EII', 'PUBLISHING'): 'CO2EII_AVAILABLE_GENERATORS',
}

# There are some tables we skip just because they are huge
# most of these are each larger than the 200 smallest tables
tables_to_skip = [
    'P5MIN_UNITSOLUTION', 
    'P5MIN_CASESOLUTION',
    'P5MIN_CONSTRAINTSOLUTION', 
    'P5MIN_LOCALPRICE', # not what people get paid. Related to electrical constraints
    'DISPATCH_LOCALPRICE', # not what people get paid. Related to electrical constraints
    'NETWORK_OUTAGEDETAIL',
    'MCC_CASESOLUTION',
    'MCC_CONSTRAINTSOLUTION',
    'DISPATCHOFFERTRK', 
    'DISPATCHCASESOLUTION',
    'P5MIN_REGIONSOLUTION', 
    'DISPATCHCONSTRAINT',
    'SETFCASREGIONRECOVERY',
    'TRADINGPRICE', # surprisingly big
]

# returns None if we can't find a table
def map_table(report_name, report_subtype) -> Optional[str]:
    candidates = metadata_df.loc[(metadata_df[report_name_column] == report_name) & (metadata_df[report_subtype_column] == report_subtype), table_name_column]
    # sometimes there's duplicates
    # but assert they're all the same answer
    tables = set(candidates)
    if len(tables) == 0 and (report_name, report_subtype) in package_exceptions:
        return package_exceptions[(report_name, report_subtype)]
        
    if len(tables) != 1:
        return None
    table = tables.pop() # choose the only one
    return table

# unit testing
assert map_table('DISPATCH', 'PRICE') == 'DISPATCHPRICE'

## Run the above code

files = [os.path.join(source_dir, p) for p in os.listdir(source_dir)]


# shuffle file order
# so that we don't do all the small files first
# then all the large files last 
# (or the other way around)
# this makes the progress bar more accurate
shuffle(files) # mutates list in place

if use_multiprocessing:
    with Pool(utils.num_cpu()) as p:
        statuses = list(tqdm(p.imap(_process_raw, files), total=len(files)))
else:
    #ret = [process_raw(file) for file in tqdm(files)]
    statuses = [_process_raw(file) for file in tqdm(files)]
assert all(s is None for s in statuses), "some files failed to be processed, to debug, run the next cell"

# Check the files which didn't download properly
# e.g. do they matter for our case?
# can you find it elsewhere?
# e.g. PUBLIC_ROOFTOP_PV_ACTUAL_MEASUREMENT_20231226053000_0000000406806404.zip was corrupted
# but can be found inside PUBLIC_ROOFTOP_PV_ACTUAL_MEASUREMENT_20231221.zip
# have a look in bad_zip_dir and bad_csv_dir for the troublesome files and error info written in adjacent text files.
[(str(s)[:100], f) for (s,f) in zip(statuses, files) if s]

files = [os.path.join(source_dir, p) for p in os.listdir(source_dir)]
[f for f in files if 'DUD' in f][-100:]
