==> 03-astral-sun-hours.ipynb <==
# Generate Sunlight data

This script generates data about how many hours of sun there are per day, per region.

We use the `astral` library for this.

For now, we're using the capital city of each region. We should come back and change this to the midpoint (by latitude) of each region. (Since most solar is not in capital cities.)

This data does not account for cloud cover.

import datetime as dt
import os

from astral.sun import sun
from astral.geocoder import database, lookup
from astral import LocationInfo
import pandas as pd
import matplotlib.pyplot as plt
import pandas as pd

base_data_dir = 'data'
output_path = os.path.join(base_data_dir, '03-sun-hours.csv')

SEC_PER_H = 60*60

regions = {
    'QLD1': 'Brisbane',
    'NSW1': 'Sydney',
    'VIC1': 'Melbourne',
    'TAS1': 'Hobart',
    'SA1': 'Adelaide'
}
regions = {r: lookup(c, database()) for (r,c) in regions.items()}
regions

data = []
d = dt.date(2009, 1, 1)

while d.year < 2025:
    for (region, city) in regions.items():
        s = sun(city.observer, date=d)
        data.append({
            'd': d,
            'regionid': region,
            'sun_hours': (s['sunrise'] - s['sunset']).total_seconds() / SEC_PER_H,
        })
    d = d + dt.timedelta(days=1)
df = pd.DataFrame(data)
df.head()

df.to_csv(output_path, index=False)
