==> 08-population-data-cleaner.ipynb <==
import pandas as pd

# Load Data
#data_dir = 'C:/Users/David/Documents/VWL/Master Toulouse/Semester 2 M1/Applied  Metrics Project/Data'
data_dir = '/home/matthew/data/'

population_raw = pd.read_csv(f'{data_dir}/population data/population-australia-raw.csv')
population_raw.head()

#First data cleaning
population = pd.concat([population_raw.iloc[:,:1],population_raw.iloc[:,-9:]],axis = 1)
population = population.drop(list(range(0, 10)))
population.columns=["Date", "NSW1", "VIC1", "QLD1", "SA1", "WA1", "TAS1", "NT1", "ACT1","AUS"]

#Include Australian Capital Territory in New South Wales
population.iloc[:, 1:] = population.iloc[:, 1:].apply(pd.to_numeric)
population['NSW1'] = population['NSW1'] + population['ACT1']
population = population.drop(['ACT1','AUS', 'NT1', 'WA1' ], axis=1)

#Transform dates to datetime format
population['Date'] = pd.to_datetime(population['Date'], format='%b-%Y')
population = population[population['Date'] >= '2008-12-01']

#Pivote the dataframe to have one column per state
population = population.melt(id_vars=['Date'], var_name='State', value_name='population')
population.head()

#Save data to CSV
population.to_csv(f'{data_dir}/08-population-australia-merged.csv', index = False)
print('All data merged and saved to CSV')
