==> 09-population-weather-merger.ipynb <==
import pandas as pd

# Load Data
#data_dir = 'C:/Users/David/Documents/VWL/Master Toulouse/Semester 2 M1/Applied  Metrics Project/Data'
data_dir = '/home/matthew/data/'

weather = pd.read_csv(f'{data_dir}/07-weather-merged.csv')
population = pd.read_csv(f'{data_dir}/08-population-australia-merged.csv')
sunshine = pd.read_csv(f'{data_dir}/07-a-sunshine-merged.csv')

# Rename column in population dataframe
population = population.rename(columns={"State": "regionid"})

# Merge data frames
merged_df = pd.merge(population, weather, on=["regionid", "Date"], how='outer')
merged_df = pd.merge(merged_df, sunshine, on=["regionid", "Date"], how='outer') 

# Sort values
merged_df = merged_df.sort_values(by=["regionid", "Date"])

# Forward fill missing values in Population column within each group
merged_df['population'] = merged_df.groupby('regionid')['population'].transform(lambda group: group.interpolate(method='linear'))

#Drop Dates before 2009 and after 2023
merged_df = merged_df.drop(merged_df[(merged_df['Date'] <= '2008-12-31') & (merged_df['Date'] >= '2024-01-01')].index)
merged_df.head()

#Save to csv
merged_df.to_csv(f'{data_dir}/09-temp-pop-merged.csv', index = False)
print("merged and saved to csv")
